+++ src/app/loadout-builder/generated-sets/GeneratedSet.tsx
@@ -123,8 +123,9 @@ export default memo(function GeneratedSet({
   const boostedStats = useMemo(
     () =>
       new Set(
-        armorStats.filter((hash) =>
-          modStatChanges[hash].breakdown?.some((change) => change.source === 'runtimeEffect')
+        armorStats.filter(
+          (hash) =>
+            modStatChanges[hash].breakdown?.some((change) => change.source === 'runtimeEffect')
         )
       ),
     [modStatChanges]