+++ src/app/records/PresentationNodeLeaf.tsx
@@ -1,5 +1,6 @@
 import { DestinyCollectibleState, DestinyRecordState } from 'bungie-api-ts/destiny2';
 import { sortBy } from 'lodash';
+import { VendorItemDisplay } from 'app/vendors/VendorItemComponent';
 import Collectible from './Collectible';
 import CollectiblesGrid from './CollectiblesGrid';
 import Craftable from './Craftable';
@@ -55,6 +56,14 @@ export default function PresentationNodeLeaf({
         <CollectiblesGrid>
           {node.craftables.map((craftable) => (
             <Craftable key={craftable.item.hash} craftable={craftable} />
+          ))}
+        </CollectiblesGrid>
+      )}
+
+      {node.plugs && node.plugs.length > 0 && (
+        <CollectiblesGrid>
+          {node.plugs.map(({ item, unlocked }) => (
+            <VendorItemDisplay key={item.index} item={item} unavailable={!unlocked} owned={false} />
           ))}
         </CollectiblesGrid>
       )}