+++ src/app/wishlists/wishlist-file.ts
@@ -99,7 +99,7 @@ function parseBlockNoteLine(blockNoteLine: string): string | undefined {
 }
 
 function getPerks(matchResults: RegExpMatchArray): Set<number> {
-  if (!matchResults.groups || matchResults.groups.itemPerks === undefined) {
+  if (matchResults.groups?.itemPerks === undefined) {
     return emptySet<number>();
   }